polish: final cosmetic pass — ruff 35 residual, 0 warnings, format clean
- Fixed 18 RUF059 (unused unpacked vars) + 2 RUF046 (unnecessary int casts)
- Fixed 2 RUF100 (unused noqa) + 1 UP035 (deprecated import)
- Suppressed scipy SparseEfficiencyWarning in pytest config
- Ruff: 35 residual (all non-actionable: E501 templates, S-series test fixtures)
- Bio mypy --strict: 0 errors (7 files)
- 1921 tests, 0 failures, 0 warnings
- Selfcheck 9/9 ALL PASS

+++ src/mycelium_fractal_net/bio/dispersal.py
@@ -78,8 +78,8 @@ class SporeDispersalEngine:
                     r = cfg.r_min * u ** (-1.0 / cfg.alpha_levy)
                     r = min(r, N * 0.7)
                     angle = rng.uniform(0, 2 * np.pi)
-                    ti = (coord[0] + int(round(r * np.sin(angle)))) % N
-                    tj = (coord[1] + int(round(r * np.cos(angle)))) % N
+                    ti = (coord[0] + round(r * np.sin(angle))) % N
+                    tj = (coord[1] + round(r * np.cos(angle))) % N
                     spore_air_new[ti, tj] += 1.0
                     new_events += 1
